Change float->double for renderable object constructors (issue #5)

+++ samples/mesh_rendering/main.cpp
@@ -22,18 +22,18 @@ int main(void)
    *******************/
 
    // Quads are built from a position, a normal, and width/height
-   float quadCenter1[3] = {-0.5, 0.0, 2.0 };
-   float quadNormal1[3] = { 0.0, 0.0,-1.0 };
+   double quadCenter1[3] = {-0.5, 0.0, 2.0 };
+   double quadNormal1[3] = { 0.0, 0.0,-1.0 };
 
    ar::Quad quad1 = ar::Quad(quadCenter1, quadNormal1, 1.0f, 1.5f, ar::Color( 0.5, 0.5, 0.25, 1.0 ));
 
-  float quadCenter2[3] = { 0.5, 0.0, 2.0 };
-  float quadNormal2[3] = { 0.0, 0.0,-1.0 };
+  double quadCenter2[3] = { 0.5, 0.0, 2.0 };
+  double quadNormal2[3] = { 0.0, 0.0,-1.0 };
 
   ar::Quad quad2 = ar::Quad(quadCenter2, quadNormal2, 0.5f, 0.5f, ar::Color( 0.5, 0.0, 0.5, 1.0 ));
 
    // Triangles are specified by three points in space
-  float triVertices1[9] = {
+  double triVertices1[9] = {
       0.0, 0.5,  3.0,
       0.5, 0.25, 2.0,
       0.0, 0.0,  1.0
@@ -50,22 +50,22 @@ int main(void)
   ar::Triangle tri2 = ar::Triangle(triVertices2[0], triVertices2[1], triVertices2[2], ar::Color( 1.0, 0.5, 0.25, 0.25 ));
 
    // A sphere is specified with just a position and a radius
-  float sphereCenter[3] = { 0.0, 0.0, 4.0 };
+  double sphereCenter[3] = { 0.0, 0.0, 4.0 };
 
   ar::Sphere sphere = ar::Sphere(sphereCenter, 2.0, ar::Color( 0.6, 0.35, 0.2, 0.7 ));
 
    // Capsules need two positions (one for each end) and a radius
-  float capsuleCenter1[3] = { 2.5, 2.0, 6.0 };
-  float capsuleCenter2[3] = { 2.0,-1.0, 5.0 };
+  double capsuleCenter1[3] = { 2.5, 2.0, 6.0 };
+  double capsuleCenter2[3] = { 2.0,-1.0, 5.0 };
 
   ar::Capsule capsule = ar::Capsule(capsuleCenter1, capsuleCenter2, 1.0, ar::Color( 0.1, 0.8, 0.3, 1.0 ));
 
    // Cubes just need a position and a size
-  float cubeCenter[3] = { -2.5, 2.0, 6.0 };
+  double cubeCenter[3] = { -2.5, 2.0, 6.0 };
   ar::Cube cube = ar::Cube(cubeCenter, 1.5, ar::Color( 1.0, 1.0, 0.0 ));
 
    // Boxes are like cubes, but can have different dimensions along each axis
-  float boxCenter[3] = { 0, -1.5, 5.0 };
+  double boxCenter[3] = { 0, -1.5, 5.0 };
   ar::Box box = ar::Box(boxCenter, 1.5, 3.5, 0.25, ar::Color( 0.5, 0.2, 0.8));
 
   // tell the visualizer about our fancy new shapes
